+++ src/modules/edc-demo/components/asset-editor-dialog/data-address-type.ts
@@ -1,1 +1,0 @@
-export type DataAddressType = 'Custom-Datasource-Json' | 'Http';