+++ src/modules/edc-demo/components/contract-agreement-transfer-dialog/contract-agreement-transfer-dialog-result.ts
@@ -1,0 +1,6 @@
+import {TransferId} from '../../../edc-dmgmt-client';
+
+export interface ContractAgreementTransferDialogResult {
+  transferProcessId: TransferId;
+  contractId: string;
+}