+++ src/modules/edc-demo/components/data-address-type-select/data-address-type-select-mode.ts
@@ -1,0 +1,1 @@
+export type DataAddressTypeSelectMode = 'Datasource' | 'Datasink';