+++ src/modules/edc-demo/components/data-address-type-select/data-address-type.ts
@@ -1,0 +1,1 @@
+export type DataAddressType = 'Custom-Data-Address-Json' | 'Http';